requirements.txt
# Kafka
kafka-python==2.0.2
confluent-kafka==2.0.2

# Spark
pyspark==3.3.1
py4j==0.10.9.5

# Data Quality
great-expectations==0.15.42
pandas==1.5.3
numpy==1.24.3

# Database
psycopg2-binary==2.9.5
sqlalchemy==2.0.7

# Configuration and utilities
pyyaml==6.0
python-dotenv==1.0.0

# Monitoring and logging
prometheus-client==0.16.0
grafana-client==2.2.0
python-json-logger==2.0.7

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Web dashboard (optional)
flask==2.3.2
dash==2.9.3 